refactor(core): consolidate unlocked vault session service
