+++ packages/core/src/__tests__/fixtures/unlock-vault.ts
@@ -2,7 +2,6 @@ import { vi } from "vitest";
 import type { DeviceAccessMaterial } from "../../domain/device/device-access-material";
 import type { VaultSnapshot } from "../../domain/snapshot/vault-snapshot";
 import { UnlockVaultUseCase } from "../../use-cases/vault-lifecycle/unlock-vault";
-import { CommitUnlockedVaultSessionService } from "../../application/vault-session/commit-unlocked-vault-session.service";
 import { createCoreTestPorts } from "./ports";
 import { createCoreTestValues } from "./values";
 
@@ -58,10 +57,6 @@ export function createUnlockVaultTestContext() {
   ports.saved.deviceAccessMaterial = deviceAccessMaterial;
   ports.saved.vaultSnapshot = vaultSnapshot;
 
-  const commitUnlockedVaultSession = new CommitUnlockedVaultSessionService(
-    ports.sessionServices.saveUnlockedVaultSession,
-    ports.sessionServices.removeUnlockedVaultSession,
-  );
   const useCase = new UnlockVaultUseCase(
     ports.clock,
     ports.crypto,
@@ -69,8 +64,7 @@ export function createUnlockVaultTestContext() {
     ports.scheduledTasks,
     ports.vaultLocalRepository,
     ports.vaultLockTasks,
-    ports.sessionServices.assertUnlockedVaultSessionCanActivate,
-    commitUnlockedVaultSession,
+    ports.sessionServices.unlockedVaultSession,
   );
 
   return {
