+++ packages/core/src/application/vault-session/index.ts
@@ -1,7 +1,1 @@
-export * from "./assert-unlocked-vault-session-can-activate.service";
-export * from "./commit-unlocked-vault-session.service";
-export * from "./get-unlocked-vault-session.service";
-export * from "./protect-unlocked-vault-session.service";
-export * from "./remove-unlocked-vault-session.service";
-export * from "./restore-unlocked-vault-session.service";
-export * from "./save-unlocked-vault-session.service";
+export * from "./unlocked-vault-session.service";
