+++ packages/core/src/use-cases/vault-entries/get-entry-password.ts
@@ -1,6 +1,6 @@
 import { PasswordEntryNotFoundError } from "../../application/errors/vault-entry.errors";
 import { VaultMustBeUnlockedError } from "../../application/errors/vault-session.errors";
-import type { GetUnlockedVaultSessionService } from "../../application/vault-session/get-unlocked-vault-session.service";
+import type { UnlockedVaultSessionService } from "../../application/vault-session/unlocked-vault-session.service";
 
 export type GetEntryPasswordCommandParams = {
   vaultId: string;
@@ -12,16 +12,16 @@ export type GetEntryPasswordResult = {
 };
 
 export class GetEntryPasswordUseCase {
-  private readonly getUnlockedVaultSession: GetUnlockedVaultSessionService;
+  private readonly unlockedVaultSession: UnlockedVaultSessionService;
 
-  constructor(getUnlockedVaultSession: GetUnlockedVaultSessionService) {
-    this.getUnlockedVaultSession = getUnlockedVaultSession;
+  constructor(unlockedVaultSession: UnlockedVaultSessionService) {
+    this.unlockedVaultSession = unlockedVaultSession;
   }
 
   async execute(
     params: GetEntryPasswordCommandParams,
   ): Promise<GetEntryPasswordResult> {
-    const unlockedVaultSession = await this.getUnlockedVaultSession.get();
+    const unlockedVaultSession = await this.unlockedVaultSession.get();
     const unlockedVault = unlockedVaultSession?.unlockedVault;
 
     if (unlockedVault?.vaultId !== params.vaultId) {
