+++ packages/core/src/use-cases/vault-entries/search-entries.ts
@@ -5,7 +5,7 @@ import { entryMatchesSearchQuery } from "../../domain/entry/search-entry-query.u
 import type { VisiblePasswordEntryFields } from "../../domain/entry/password-entry.type";
 import { InvalidSearchEntryQueryError } from "../../application/errors/vault-entry.errors";
 import { VaultMustBeUnlockedError } from "../../application/errors/vault-session.errors";
-import type { GetUnlockedVaultSessionService } from "../../application/vault-session/get-unlocked-vault-session.service";
+import type { UnlockedVaultSessionService } from "../../application/vault-session/unlocked-vault-session.service";
 
 export type SearchEntriesCommandParams = {
   vaultId: string;
@@ -17,16 +17,16 @@ export type SearchEntriesResult = {
 };
 
 export class SearchEntriesUseCase {
-  private readonly getUnlockedVaultSession: GetUnlockedVaultSessionService;
+  private readonly unlockedVaultSession: UnlockedVaultSessionService;
 
-  constructor(getUnlockedVaultSession: GetUnlockedVaultSessionService) {
-    this.getUnlockedVaultSession = getUnlockedVaultSession;
+  constructor(unlockedVaultSession: UnlockedVaultSessionService) {
+    this.unlockedVaultSession = unlockedVaultSession;
   }
 
   async execute(
     params: SearchEntriesCommandParams,
   ): Promise<SearchEntriesResult> {
-    const unlockedVaultSession = await this.getUnlockedVaultSession.get();
+    const unlockedVaultSession = await this.unlockedVaultSession.get();
     const unlockedVault = unlockedVaultSession?.unlockedVault;
 
     if (unlockedVault?.vaultId !== params.vaultId) {
