+++ packages/core/src/use-cases/vault-lifecycle/lock-vault.ts
@@ -2,7 +2,7 @@ import type { ClipboardClearTaskRepositoryPort } from "../../ports/clipboard/cli
 import type { ScheduledTaskPort } from "../../ports/system/scheduled-task.port";
 import type { VaultLockTaskRepositoryPort } from "../../ports/vault/vault-lock-task-repository.port";
 import type { ClearClipboardTaskUseCase } from "../clipboard/clear-clipboard-task";
-import type { RemoveUnlockedVaultSessionService } from "../../application/vault-session/remove-unlocked-vault-session.service";
+import type { UnlockedVaultSessionService } from "../../application/vault-session/unlocked-vault-session.service";
 
 export type LockVaultCommandParams = {
   actionId?: string;
@@ -13,20 +13,20 @@ export class LockVaultUseCase {
   private readonly clipboardClearTasks: ClipboardClearTaskRepositoryPort;
   private readonly scheduledTasks: ScheduledTaskPort;
   private readonly vaultLockTasks: VaultLockTaskRepositoryPort;
-  private readonly removeUnlockedVaultSession: RemoveUnlockedVaultSessionService;
+  private readonly unlockedVaultSession: UnlockedVaultSessionService;
 
   constructor(
     clearClipboardTask: ClearClipboardTaskUseCase,
     clipboardClearTasks: ClipboardClearTaskRepositoryPort,
     scheduledTasks: ScheduledTaskPort,
     vaultLockTasks: VaultLockTaskRepositoryPort,
-    removeUnlockedVaultSession: RemoveUnlockedVaultSessionService,
+    unlockedVaultSession: UnlockedVaultSessionService,
   ) {
     this.clearClipboardTask = clearClipboardTask;
     this.clipboardClearTasks = clipboardClearTasks;
     this.scheduledTasks = scheduledTasks;
     this.vaultLockTasks = vaultLockTasks;
-    this.removeUnlockedVaultSession = removeUnlockedVaultSession;
+    this.unlockedVaultSession = unlockedVaultSession;
   }
 
   async execute(params: LockVaultCommandParams = {}): Promise<void> {
@@ -82,7 +82,7 @@ export class LockVaultUseCase {
       executionError = error;
     } finally {
       try {
-        await this.removeUnlockedVaultSession.remove();
+        await this.unlockedVaultSession.remove();
       } catch (error) {
         if (executionError === undefined) {
           sessionRemovalError = error;
